old-tests/submission/Microsoft/async/async_004.htm

<!DOCTYPE html>
<html>
    <head>
        <title>An async script executes as soon as possible after a download is complete</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="This test ensures an async script executes as soon as possible after a download is complete." />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://dev.w3.org/html5/spec/scripting-1.html#attr-script-async"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">        
        
  
        var t = async_test("async script executes as soon as possible after a download is complete");
  
        function timeout()
        {
            t.step(function(){ assert_equals(document.getElementById("testresult").innerHTML, "21")});
            t.done();
        }
  
        var timer = setTimeout(timeout, 5000);
        
        // This test uses a script (log.py) to change script order execution
        // The php script takes two parameters 
        //  * The first is for a the delay (sec=1 == sleep(1) in php)
        //  * The second is passed as a parameter to 'log()'
        //
        // The log function just appends to the testresult tag.
        //
        // In this case the log.py script tags have the 'async' attribute so the parser doesn't block execution
        // Because of the sleep in the php file the second script below executes before the first
        // 
        //  Thus the assert above is to check that the innerHTML == 21
        //
        
        function log(text) 
        {
            var textNode = document.createTextNode(text);
            document.getElementById("testresult").appendChild(textNode);
        }
        
        </script>
        
        <span style="display : none" id="testresult"></span>

        <script src="log.py?sec=3&id=1" async></script>
        <script src="log.py?sec=1&id=2" async></script>
    </body>
</html>
